backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
python-dotenv==1.0.1
pydantic==2.9.2
langchain==0.3.1
langchain-openai==0.2.1
langchain-anthropic==0.2.3
langchain-core==0.3.9
langgraph==0.2.28
crewai==0.80.0
crewai-tools==0.14.0
httpx==0.27.2
